Revamped $state.go to go to availability views from study detail create/update
Used $state.go for edit availability to pass study's availability in $stateParams, as querying for the information broke the calendar (the calendar tried to get Date Objects faster than the asynchronous http.get would return.
Still also need to get it so that in edit the select boxes show what's already chosen. The information is there and right, but need to reflect what's there in the selector.